auth

+++ pao_project_MML/app/Http/Controllers/ProductsController.php
@@ -2,12 +2,18 @@
 
 namespace App\Http\Controllers;
 
+use Auth;
 use App\Products;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
 class ProductsController extends Controller
 {
+	public function __construct()
+    {
+        $this->middleware('auth');
+    }
+	
 	public function index()
 	{
 		$products = Products::all();
@@ -16,15 +22,15 @@ class ProductsController extends Controller
 
     public function edit($id)
     {
+		if(Auth::user()->rank == 0)
+		{
+			return redirect()->to('/');
+		}
+		
         $products = Products::where('id', $id)->get();
         return view('products.edit', compact('products'));
     }
 	
-	/**
-     * 
-     * Show method in ProductsController
-     * 
-     */
 	public function product($products)
 	{
 		$product = Products::where('id', $products)->get();
@@ -34,11 +40,21 @@ class ProductsController extends Controller
 
     public function newProduct()
     {
+		if(Auth::user()->rank == 0)
+		{
+			return redirect()->to('/');
+		}
+		
         return view ('products.add');
     }
 
     public function add(Request $request)
     {
+		if(Auth::user()->rank == 0)
+		{
+			return redirect()->to('/');
+		}
+		
        $newProduct = new Products();
 
         $newProduct->name = $request->name;
@@ -55,6 +71,11 @@ class ProductsController extends Controller
 	
 	public function deleteProduct(Request $request, $product)
     {
+		if(Auth::user()->rank == 0)
+		{
+			return redirect()->to('/');
+		}
+		
 		$product = Products::find($product);
 		$product->delete($request->all());
 		
@@ -63,6 +84,11 @@ class ProductsController extends Controller
 
     public function editProduct(Request $request, $product)
     {
+		if(Auth::user()->rank == 0)
+		{
+			return redirect()->to('/');
+		}
+		
         $product = Products::find($product);
         $product->update($request->all());
